requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backcall==0.2.0
black==21.6b0
click==8.0.1
colorama==0.4.4
decorator==5.0.9
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
Faker==8.8.2
flake8==3.9.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
marshmallow==3.12.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pep8-naming==0.11.1
pickleshare==0.7.5
prompt-toolkit==3.0.19
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
webargs==8.0.0
